Float to integer for box vertices

+++ src/me/geakstr/voxel/model/AABB.java
@@ -45,23 +45,23 @@ public class AABB {
 
     public static enum SIDE {
         FRONT {
-            public float[] vertices_coords() {
+            public int[] vertices_coords() {
                 return Arrays.copyOf(front_side_vertices_coords, side_vertices_size);
             }
 
-            public float[] texture_coords(int u, int v) {
-                return new float[]{0, 1, u, 1, 0, 0, u, 0};
+            public int[] texture_coords(int u, int v) {
+                return new int[]{0, 1, u, 1, 0, 0, u, 0};
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(vertices_coords(), x_pos, y_pos, z_pos);
             }
 
-            public float[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
+            public int[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
                 x_expand--;
                 y_expand--;
                 z_expand--;
-                float[] side = translate(x_pos, y_pos, z_pos);
+                int[] side = translate(x_pos, y_pos, z_pos);
 
                 side[3] += x_expand;
                 side[9] += x_expand;
@@ -78,23 +78,23 @@ public class AABB {
             }
         },
         BACK {
-            public float[] vertices_coords() {
+            public int[] vertices_coords() {
                 return Arrays.copyOf(back_side_vertices_coords, side_vertices_size);
             }
 
-            public float[] texture_coords(int u, int v) {
-                return new float[]{0, 1, u, 1, 0, 0, u, 0};
+            public int[] texture_coords(int u, int v) {
+                return new int[]{0, 1, u, 1, 0, 0, u, 0};
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(vertices_coords(), x_pos, y_pos, z_pos);
             }
 
-            public float[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
+            public int[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
                 x_expand--;
                 y_expand--;
                 z_expand--;
-                float[] side = translate(x_pos, y_pos, z_pos);
+                int[] side = translate(x_pos, y_pos, z_pos);
 
                 side[0] += x_expand;
                 side[6] += x_expand;
@@ -106,23 +106,23 @@ public class AABB {
             }
         },
         LEFT {
-            public float[] vertices_coords() {
+            public int[] vertices_coords() {
                 return Arrays.copyOf(left_side_vertices_coords, side_vertices_size);
             }
 
-            public float[] texture_coords(int u, int v) {
-                return new float[]{v, 1, 0, 1, v, 0, 0, 0};
+            public int[] texture_coords(int u, int v) {
+                return new int[]{v, 1, 0, 1, v, 0, 0, 0};
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(vertices_coords(), x_pos, y_pos, z_pos);
             }
 
-            public float[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
+            public int[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
                 x_expand--;
                 y_expand--;
                 z_expand--;
-                float[] side = translate(x_pos, y_pos, z_pos);
+                int[] side = translate(x_pos, y_pos, z_pos);
 
                 side[7] += y_expand;
                 side[10] += y_expand;
@@ -134,23 +134,23 @@ public class AABB {
             }
         },
         RIGHT {
-            public float[] vertices_coords() {
+            public int[] vertices_coords() {
                 return Arrays.copyOf(right_side_vertices_coords, side_vertices_size);
             }
 
-            public float[] texture_coords(int u, int v) {
-                return new float[]{v, 1, 0, 1, v, 0, 0, 0};
+            public int[] texture_coords(int u, int v) {
+                return new int[]{v, 1, 0, 1, v, 0, 0, 0};
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(vertices_coords(), x_pos, y_pos, z_pos);
             }
 
-            public float[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
+            public int[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
                 x_expand--;
                 y_expand--;
                 z_expand--;
-                float[] side = translate(x_pos, y_pos, z_pos);
+                int[] side = translate(x_pos, y_pos, z_pos);
 
                 side[0] += x_expand;
                 side[3] += x_expand;
@@ -167,23 +167,23 @@ public class AABB {
             }
         },
         TOP {
-            public float[] vertices_coords() {
+            public int[] vertices_coords() {
                 return Arrays.copyOf(top_side_vertices_coords, side_vertices_size);
             }
 
-            public float[] texture_coords(int u, int v) {
-                return new float[]{0, v, u, v, 0, 0, u, 0};
+            public int[] texture_coords(int u, int v) {
+                return new int[]{0, v, u, v, 0, 0, u, 0};
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(vertices_coords(), x_pos, y_pos, z_pos);
             }
 
-            public float[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
+            public int[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
                 x_expand--;
                 y_expand--;
                 z_expand--;
-                float[] side = translate(x_pos, y_pos, z_pos);
+                int[] side = translate(x_pos, y_pos, z_pos);
 
                 side[3] += x_expand;
                 side[9] += x_expand;
@@ -200,23 +200,23 @@ public class AABB {
             }
         },
         BOTTOM {
-            public float[] vertices_coords() {
+            public int[] vertices_coords() {
                 return Arrays.copyOf(bottom_side_vertices_coords, side_vertices_size);
             }
 
-            public float[] texture_coords(int u, int v) {
-                return new float[]{0, v, u, v, 0, 0, u, 0};
+            public int[] texture_coords(int u, int v) {
+                return new int[]{0, v, u, v, 0, 0, u, 0};
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(vertices_coords(), x_pos, y_pos, z_pos);
             }
 
-            public float[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
+            public int[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
                 x_expand--;
                 y_expand--;
                 z_expand--;
-                float[] side = translate(x_pos, y_pos, z_pos);
+                int[] side = translate(x_pos, y_pos, z_pos);
 
                 side[3] += x_expand;
                 side[9] += x_expand;
@@ -228,12 +228,12 @@ public class AABB {
             }
         },
         ALL {
-            public float[] vertices_coords() {
+            public int[] vertices_coords() {
                 return Arrays.copyOf(all_sides_vertices_coords, sides_vertices_size);
             }
 
-            public float[] texture_coords(int u, int v) {
-                return new float[]{
+            public int[] texture_coords(int u, int v) {
+                return new int[]{
                         // front
                         0, 1, u, 1, 0, 0, u, 0,
                         // back
@@ -249,15 +249,15 @@ public class AABB {
                 };
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(vertices_coords(), x_pos, y_pos, z_pos);
             }
 
-            public float[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
+            public int[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand) {
                 x_expand--;
                 y_expand--;
                 z_expand--;
-                float[] side = translate(x_pos, y_pos, z_pos);
+                int[] side = translate(x_pos, y_pos, z_pos);
 
                 // front
                 side[3] += x_expand;
@@ -320,108 +320,108 @@ public class AABB {
             }
         };
 
-        public abstract float[] vertices_coords();
+        public abstract int[] vertices_coords();
 
-        public abstract float[] texture_coords(int u, int v);
+        public abstract int[] texture_coords(int u, int v);
 
-        public abstract float[] translate(int x_pos, int y_pos, int z_pos);
+        public abstract int[] translate(int x_pos, int y_pos, int z_pos);
 
-        public abstract float[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand);
+        public abstract int[] translate_and_expand(int x_pos, int y_pos, int z_pos, int x_expand, int y_expand, int z_expand);
 
         public static final SIDE[] values = values();
     }
 
     public static enum VERTEX {
         V0 {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(v0_vertex_coords, 3);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         },
         V1 {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(v1_vertex_coords, 3);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         },
         V2 {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(v2_vertex_coords, 3);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         },
         V3 {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(v3_vertex_coords, 3);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         },
         V4 {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(v4_vertex_coords, 3);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         },
         V5 {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(v5_vertex_coords, 3);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         },
         V6 {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(v6_vertex_coords, 3);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         },
         V7 {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(v7_vertex_coords, 3);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         },
         ALL {
-            public float[] coords() {
+            public int[] coords() {
                 return Arrays.copyOf(all_vertices_coords, 18);
             }
 
-            public float[] translate(int x_pos, int y_pos, int z_pos) {
+            public int[] translate(int x_pos, int y_pos, int z_pos) {
                 return AABB.translate(coords(), x_pos, y_pos, z_pos);
             }
         };
 
-        public abstract float[] coords();
+        public abstract int[] coords();
 
-        public abstract float[] translate(int x_pos, int y_pos, int z_pos);
+        public abstract int[] translate(int x_pos, int y_pos, int z_pos);
 
         public static final VERTEX[] values = values();
     }
 
-    private static float[] translate(float[] vertices_coords, int x_pos, int y_pos, int z_pos) {
+    private static int[] translate(int[] vertices_coords, int x_pos, int y_pos, int z_pos) {
         int size = vertices_coords.length;
 
         for (int i = 0; i < size; i += 3) {
@@ -433,7 +433,7 @@ public class AABB {
         return vertices_coords;
     }
 
-    private static final float[] all_sides_vertices_coords = new float[]{
+    private static final int[] all_sides_vertices_coords = new int[]{
             // front
             0, 0, 1, // v1
             1, 0, 1, // v2
@@ -471,7 +471,7 @@ public class AABB {
             1, 0, 1, // v2
     };
 
-    private static final float[] front_side_vertices_coords = new float[]{
+    private static final int[] front_side_vertices_coords = new int[]{
             0, 0, 1, // v1
             1, 0, 1, // v2
             0, 1, 1, // v0
@@ -482,7 +482,7 @@ public class AABB {
             1, 2, 0, 3
     };
 
-    private static final float[] back_side_vertices_coords = new float[]{
+    private static final int[] back_side_vertices_coords = new int[]{
             1, 0, 0, // v6
             0, 0, 0, // v7
             1, 1, 0, // v5
@@ -493,7 +493,7 @@ public class AABB {
             6, 7, 5, 4
     };
 
-    private static final float[] left_side_vertices_coords = new float[]{
+    private static final int[] left_side_vertices_coords = new int[]{
             0, 0, 0, // v7
             0, 0, 1, // v1
             0, 1, 0, // v4
@@ -504,7 +504,7 @@ public class AABB {
             7, 1, 4, 0
     };
 
-    private static final float[] right_side_vertices_coords = new float[]{
+    private static final int[] right_side_vertices_coords = new int[]{
             1, 0, 1, // v2
             1, 0, 0, // v6
             1, 1, 1, // v3
@@ -515,7 +515,7 @@ public class AABB {
             2, 6, 3, 5
     };
 
-    private static final float[] top_side_vertices_coords = new float[]{
+    private static final int[] top_side_vertices_coords = new int[]{
             0, 1, 1, // v0
             1, 1, 1, // v3
             0, 1, 0, // v4
@@ -526,7 +526,7 @@ public class AABB {
             0, 3, 4, 5
     };
 
-    private static final float[] bottom_side_vertices_coords = new float[]{
+    private static final int[] bottom_side_vertices_coords = new int[]{
             0, 0, 0, // v7
             1, 0, 0, // v6
             0, 0, 1, // v1
@@ -537,7 +537,7 @@ public class AABB {
             7, 6, 1, 2
     };
 
-    private static final float[] all_vertices_coords = new float[]{
+    private static final int[] all_vertices_coords = new int[]{
             0, 1, 1, // v0
             0, 0, 1, // v1
             1, 0, 1, // v2
@@ -548,35 +548,35 @@ public class AABB {
             0, 0, 0, // v7
     };
 
-    private static final float[] v0_vertex_coords = new float[]{
+    private static final int[] v0_vertex_coords = new int[]{
             0, 1, 1, // v0
     };
 
-    private static final float[] v1_vertex_coords = new float[]{
+    private static final int[] v1_vertex_coords = new int[]{
             0, 0, 1, // v1
     };
 
-    private static final float[] v2_vertex_coords = new float[]{
+    private static final int[] v2_vertex_coords = new int[]{
             1, 0, 1, // v2
     };
 
-    private static final float[] v3_vertex_coords = new float[]{
+    private static final int[] v3_vertex_coords = new int[]{
             1, 1, 1, // v3
     };
 
-    private static final float[] v4_vertex_coords = new float[]{
+    private static final int[] v4_vertex_coords = new int[]{
             0, 1, 0, // v4
     };
 
-    private static final float[] v5_vertex_coords = new float[]{
+    private static final int[] v5_vertex_coords = new int[]{
             1, 1, 0, // v5
     };
 
-    private static final float[] v6_vertex_coords = new float[]{
+    private static final int[] v6_vertex_coords = new int[]{
             1, 0, 0, // v6
     };
 
-    private static final float[] v7_vertex_coords = new float[]{
+    private static final int[] v7_vertex_coords = new int[]{
             0, 0, 0, // v7
     };
 }
